# [#3654] Synchronize on PoolSubpage head when allocate / free PoolSubp…