@normanmaurer 2015-05-27
@trustin please review

@nmittler @louiscryan @ejona86 again interesting for grpc :)