@normanmaurer 2015-05-27
Cherry-picked into 4.0 (f18990a8a507d52fc40416d169db340105b10ec0), 4.1 (ab1bb9136b384e4722feebe2f0e9377cd88fd39a) and master (3a731041b415fae6ba73615f7d0e23ffcd9a774a)